refactor(#18): ShopController Swagger 명세 인터페이스로 분리
